+++ packages/common/src/components/atoms/Button/Button.stories.tsx
@@ -1,13 +1,20 @@
-import React from 'react'
-import { Button } from './Button'
+import { Story } from '@storybook/react'
+import { Button, CustomButtonProps } from './Button'
 
 export default {
-  title: 'Button',
+  title: 'Atoms/Buttons/Button',
   component: Button,
+  argTypes: {
+    text: { control: 'text' },
+  },
 }
 
-const runAlert = () => {
-  alert('Hello!')
-}
+const runAlert = () => alert('Hello!')
 
-export const Regular = () => <Button title="Welcome" onClick={runAlert} />
+const Template: Story<CustomButtonProps> = args => <Button {...args} />
+
+export const Regular = Template.bind({})
+Regular.args = {
+  text: 'Welcome',
+  onClick: runAlert,
+}