+++ packages/common/src/components/atoms/Button/Button.test.tsx
@@ -1,10 +1,16 @@
-import { render, screen } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import { Button } from './Button'
 
 describe('Button', () => {
   test('it renders successfully', () => {
-    render(<Button title="hola" />)
-    const buttonElement = screen.getByText(/hola/i)
+    const { getByTestId } = render(<Button text="hola" />)
+    const buttonElement = getByTestId(/testButton/i)
+    expect(buttonElement).toBeInTheDocument()
+  })
+
+  test('it renders text property', () => {
+    const { getByText } = render(<Button text="hola" />)
+    const buttonElement = getByText(/hola/i)
     expect(buttonElement).toBeInTheDocument()
   })
 })