+++ packages/common/src/components/atoms/Flex/Flex.test.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+import { Flex } from './Flex'
+
+describe('Flex', () => {
+  test('it renders successfully', () => {
+    const { getByTestId } = render(<Flex />)
+    const flexElement = getByTestId(/testFlex/i)
+    expect(flexElement).toBeInTheDocument()
+  })
+})