+++ packages/common/src/index.ts
@@ -1,1 +1,2 @@
+export * from './global.d'
 export * from './components'