+++ packages/webapp-redux/src/containers/counter/Counter.styles.tsx
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled'
+
+export const CounterSpan = styled.span`
+  font-size: 78px;
+  padding-left: 16px;
+  padding-right: 16px;
+  margin-top: 2px;
+  font-family: 'Courier New', Courier, monospace;
+`
+
+export const CounterInput = styled.input`
+  font-size: 24px;
+  padding: 2px;
+  width: 64px;
+  text-align: center;
+  margin-right: 4px;
+  height: 30px;
+`