+++ packages/webapp-redux/src/index.tsx
@@ -1,0 +1,30 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { Provider } from 'react-redux'
+import { Global, css } from '@emotion/react'
+
+import { store } from './app/store'
+import { App } from './App'
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Global
+      styles={css`
+        body {
+          margin: 0;
+          font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell',
+            'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+          -webkit-font-smoothing: antialiased;
+          -moz-osx-font-smoothing: grayscale;
+        }
+        code {
+          font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+        }
+      `}
+    />
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </React.StrictMode>,
+  document.getElementById('root'),
+)