+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/sauldeleon/monorepo-starter/compare/v1.0.1...v1.1.0) (2021-06-06)
+
+
+### Features
+
+* **common:** add new atoms and update depencencies ([9c12601](https://github.com/sauldeleon/monorepo-starter/commit/9c12601e21e379a6e666e035472ba376daa2a7c3))
+* **webapp-redux:** add new project ([16bcb5f](https://github.com/sauldeleon/monorepo-starter/commit/16bcb5f203e5a67a4419a883399ed1127a602934))
+
+
+### Documentation updates
+
+* update docs ([d26329c](https://github.com/sauldeleon/monorepo-starter/commit/d26329c271cca340b0c82e5708963e20408baf48))
+
+
+### Chore
+
+* **webapp:** update dependencies ([260beb4](https://github.com/sauldeleon/monorepo-starter/commit/260beb432420b0f7e3a76a0c886cd50d4cd09e11))
+* update dependencies and improve config ([7194d43](https://github.com/sauldeleon/monorepo-starter/commit/7194d43c81383b1e5a00e6b0252f5f9df0ad1cdc))
+
 ### 1.0.1 (2021-05-14)